finally seperated out the models, db seems to need to be in its own file
